requirements.txt
aiobotocore==2.13.1
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aioitertools==0.11.0
aiosignal==1.3.1
anyio==4.4.0
attrs==24.1.0
botocore==1.34.131
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
condastats==0.2.1
cramjam==2.8.3
cryptography==43.0.0
dask==2024.7.1
dask-expr==1.1.9
DataProperty==1.0.1
dominate==2.9.1
et-xmlfile==1.1.0
evdspy==1.1.39
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
iniconfig==2.0.0
jmespath==1.0.1
locket==1.0.0
markdown-it-py==3.0.0
mbstrdecoder==1.1.3
mdurl==0.1.2
multidict==6.0.5
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
partd==1.4.2
pathvalidate==3.2.0
platformdirs==4.2.2
pluggy==1.5.0
prettytable==3.10.2
pyarrow==17.0.0
pycparser==2.22
Pygments==2.18.0
pypistats==1.6.0
pytablewriter==1.2.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-snappy==0.7.2
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
s3fs==2024.6.1
six==1.16.0
sniffio==1.3.1
tabledata==1.3.3
tcolorpy==0.1.6
termcolor==2.4.0
text-unidecode==1.3
toolz==0.12.1
typepy==1.3.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
